feat(scripted-tool): add ToolDef extension
